Time: 11 ms (14.99%) | Memory: 8.6 MB (14.27%) - LeetSync

+++ 69-sqrtx/sqrtx.cpp
@@ -3,7 +3,7 @@ public:
     int mySqrt(int x) { 
         if (x==0) return 0;
         if (x==1) return 1;
-        for (long int i=1;i<=(x/2)+1;i++){
+        for (long long int i=1;i<=(x/2)+1;i++){
             if (i*i>x){
                 return i-1;
             }
